utils_py/pyproject.toml
[project]
authors = [
  {name = "Salvatore D'angelo", email = "sasadangelo@gmail.com"},
]
description = "Utility scripts for flashcards app"
name = "flashcards-utils"
version = "0.1.0"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
python-dotenv = "^1.1.1"
gtts = "^2.5.4"
diffusers = "^0.34.0"
torch = "^2.7.1"
transformers = "^4.53.1"
pillow = "^11.3.0"
